id INT to LONG was changed

+++ src/main/java/com/example/AdminAccounts.java
@@ -19,17 +19,17 @@ public class AdminAccounts {
     }
 
 
-    private  int id;
+    private Long id;
 
     private String password;
     private String login;
 
     private String username;
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
     public String getPassword() {
@@ -101,7 +101,7 @@ public class AdminAccounts {
         }
     }
 
-    public void deleteAdminAccount(int id) throws SQLException {
+    public void deleteAdminAccount(Long id) throws SQLException {
         try {
             Statement stmt = connection.createStatement();
